Bridge PReLU/PReLUScalar, Quantize/Dequantize, Softmax, FullyConnected/BatchMatMul; improve perf-stream reporting
Extends the perf-stream hardware bridge (generated_test_bridge.py,
adapter_specs.py, benchmark_server_session.c/h, kernel_registry.yaml) to
cover the remaining requested operator families end-to-end on real
Apollo510/Cortex-M55 hardware:

- ActivationFunctions PReLU / PReLUScalar
- QuantizationFunctions Quantize / Dequantize
- SoftmaxFunctions Softmax
- FullyConnectedFunctions FullyConnected / BatchMatMul

All hardware-verified on board 1160002276. Full cross-family regression
sweep: 285-286/291-292 passing, with only 6 pre-existing failures
(leaky_relu_*_s8, hard_swish_compat_nhwc_s8, depthwise_conv_*_s8) that are
genuine 1-2 LSB MVE-vs-scalar rounding divergences inside shared CMSIS-NN
kernel source, unrelated to this change (root-cause analysis documented in
docs/perf-stream-expansion-progress.md).

Also improves perf-stream CLI/script readability and reporting:

- cli.py: per-case output now grouped by operator family with colorized
  PASS/FAIL status, aligned columns, and a trailing pass/fail summary with
  a highlighted list of failing case IDs.
- cli.py: skipped-test reporting now groups by deduplicated reason instead
  of repeating an identical, ever-growing "bridged today" family list per
  case (217 near-duplicate lines collapse to ~50 counted groups).
- run_hardware_perf_suite.sh: replaced the raw full-CSV dump with a
  colorized pass/fail summary highlighting failing cases.

+++ helia_core_tester/perf_stream/session.py
@@ -123,7 +123,9 @@ class HostSession:
                     raise RuntimeError("Received OUTPUT_END without an active case.")
                 bundle = case_map[current_case_id]
                 expected_output = blob_numpy(bundle.expected_output)
-                actual = np.frombuffer(bytes(actual_output_bytes), dtype=np.int8).reshape(expected_output.shape)
+                actual = np.frombuffer(bytes(actual_output_bytes), dtype=expected_output.dtype).reshape(
+                    expected_output.shape
+                )
                 comparison_result = compare_output(actual, expected_output, bundle.comparison)
                 writer = ByteWriter()
                 writer.u8(1 if comparison_result.passed else 0)
